Enable `package-lock.json` to make builds reproducible
Locking revisions for all dependencies is required to make the build
process actually use the same code even if performed later, after some
dependencies had newer releases published.  The Iosevka maintainer
apparently has a different opinion on this, therefore the upstream code
has `package-lock=false` in `.npmrc`; change that setting.  In addition,
the `package-lock.json` file was listed in `.gitignore`; remove that
line, so that the lock file could be committed normally.