Add locked revisions for all dependencies
Add the `package-lock.json` file which specifies locked revisions for
all dependencies.  This file can be used by `npm ci` to get the exact
revisions of dependencies without picking up any updates, so that the
build results would be reproducible.

Before building a new font release, the `package-lock.json` file needs
to be updated, so that the font files would be built using the intended
revisions of various packages.  The following commands should be used:

    rm -rf package-lock.json node_modules
    npm install
    git add package-lock.json

(Commands like `npm update` or `npm update --depth 9999` do not give the
same results and should not be used.)