Add a simple YAML to TOML converter
YAML supports anchors, aliases and the “merge key” feature, which can be
used to avoid duplicate content that is usually present in the TOML
representation of build plans.  Modules for reading YAML and writing
TOML are already in the dependency tree, so adding a simple YAML to TOML
converter is easy.