Add a Makefile to build release archives
The Iosevka build system does not support building release archives for
custom fonts, and the way it builds these archives is not really good
anyway (it makes ZIP files, which provide suboptimal compression).
Handle this part with a Makefile instead, and create *.tar.xz archives.

The Makefile also handles renaming of Cosevka Basic TTF files, using
features of GNU make and GNU tar.

Dependencies for `node_modules` are handled in a way that makes it
possible to run `npm ci` before invoking the Makefile and still get the
proper results; this is required to get usable log output for GitHub
workflows (the Iosevka build system requires faking a terminal to get
any progress output, but this also enables the progress output in `npm`,
which looks really bad when not on a real terminal).